Refactor onboarding flow and remove temporary registration status
